Initial commit of completed work.
A project for my Graduate CS AI course in-which we must implement search algorithms for a Pacman game. The algorithms are: BFS, DFS, A*, and UCS. The A* had 2 aspects to it, the first, having a heuristic of 0, then again with creating a heuristic that could solve the problem of "trickySearch" (found in the commands file) in under 7,000 nodes explored.